refactor(ui): split main layout into smaller components
