+++ src/routes/(customer)/+page.server.ts
@@ -1,11 +1,12 @@
 import { db } from '$lib/server/db';
 import { events } from '$lib/server/db/schema';
-import { eq, ilike, and, desc } from 'drizzle-orm';
+import { and, count, desc, eq, ilike } from 'drizzle-orm';
 import type { PageServerLoad } from './$types';
 
 export const load: PageServerLoad = async ({ url }) => {
   // 1. Nhận từ khóa tìm kiếm và trang hiện tại từ URL (VD: ?q=rock&page=1)
   const searchQuery = url.searchParams.get('q') || '';
+  const escapedQuery = searchQuery.replace(/[%_]/g, '\\$&');
   const page = Number(url.searchParams.get('page')) || 1;
   const limit = 8; // Hiển thị 8 sự kiện 1 trang
   const offset = (page - 1) * limit;
@@ -15,7 +16,7 @@ export const load: PageServerLoad = async ({ url }) => {
 
   // Nếu user có gõ tìm kiếm, thêm điều kiện tìm theo tiêu đề (ilike: không phân biệt hoa thường)
   if (searchQuery) {
-    conditions.push(ilike(events.title, `%${searchQuery}%`));
+    conditions.push(ilike(events.title, `%${escapedQuery}%`));
   }
 
   // 3. Query Database
@@ -29,9 +30,13 @@ export const load: PageServerLoad = async ({ url }) => {
 
   // 4. Đếm tổng số sự kiện để làm phân trang (Pagination)
   // Tạm thời lấy length để làm demo nhanh, dự án thật sẽ dùng hàm COUNT() của DB
-  const totalEvents = eventList.length;
-  const totalPages = Math.ceil(totalEvents / limit) || 1;
+  const totalEvents = await db
+    .select({ count: count() })
+    .from(events)
+    .where(and(...conditions))
+    .then((result) => result[0]?.count || 0);
 
+  const totalPages = Math.ceil(totalEvents / limit);
   // 5. Trả dữ liệu về cho Frontend (+page.svelte)
   return {
     events: eventList,
